requirements.txt
cycler==0.10.0
jieba==0.39
kiwisolver==1.1.0
matplotlib==2.2.4
numpy==1.17.0
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
wordcloud==1.5.0
